requirements.txt
# NeoSignal dependencies
requests>=2.32.0
fpdf2>=2.8.0

# Dev / test
pytest>=8.2.0
pytest-mock>=3.14.0
pylint>=3.2.0
